- refactor: 서비스 상태 처리 enum 기반으로 정리 - body: - status/orderType 문자열 비교 제거 -> enum 변환(fromRequired/fromNullableParam) - PaymentService 승인/실패/만료 처리 로직 가독성 개선 - OrderQueryService Status 파라미터 검증/정규화 적용

+++ src/main/java/com/jkomerce/store/dto/BoardDTO.java
@@ -9,8 +9,8 @@ import java.time.LocalDateTime;
 @Data
 public class BoardDTO {
 
-    private int id;
-    private int userId;
+    private Long id;
+    private Integer userId;
     private String userName;
     private String title;
     private String content;
